ai-twitter-marketing-agent/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
tweepy==4.14.0
anthropic==0.43.0
apscheduler==3.10.4
jinja2==3.1.4
python-dotenv==1.0.1
aiosqlite==0.20.0
